sc_utils/requirements.txt
# Web3 and Ethereum
web3==6.11.3
eth-account==0.10.0
eth-utils==4.0.0
eth-abi==4.2.1
eth-typing==3.5.2

# Solidity compilation
py-solc-x==2.0.2

# Environment and configuration
python-dotenv==1.0.0

# HTTP requests for RPC and API calls
requests==2.31.0
aiohttp==3.9.1

# Utilities
click==8.1.7  # CLI interface
colorama==0.4.6  # Colored terminal output
tabulate==0.9.0  # Table formatting

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: for hardware wallet support in future
# ledgereth==0.9.1
# trezor==0.13.8